Refactor backtest into module
